Update ex6.asm
Hello there! Thanks for the lessons. I was watching 'Intro to x86 Assembly Language (Part 3)' while trying out everything on my terminal open on the right. Found something weird running the code at 9:50. It printed a highlighted `%` at the end of the message on my system. So, I increased the stack size and added an `0x0a` at the end of the string and added `1` to the contents of the `edx` register. It seemed that the problem was solved. I tried to see the exit status of the code at the end of part 1 by moving `len` to the `ebx` register before exiting. Guess what? It added a `1` to the string length as well! It seems like we need to include the `0x0a` byte when we address the length of the string. Please feel free to correct me if I went wrong anywhere. Thank you. Stay safe. :)